searchFilte debug done
