+++ src/layouts/HeaderRightBox.jsx
@@ -1,13 +1,12 @@
-import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useRecoilState, useResetRecoilState } from 'recoil';
 
 import { Icon } from '@/components/common/Icon/Icon.jsx';
 import { localstorageKey } from '@/constant/index.js';
-import { itemListState } from '@/recoil/atoms/itemListState.js';
 import { customerState, loginState, userState } from '@/recoil/atoms/userState.js';
 import { removeItem } from '@/utils/localstorage.js';
 import * as S from './Header.Styles.jsx';
+import SearchBar from './SearchBar.jsx';
 
 const HeaderRightBox = () => {
   // recoil의 상태 정보 불러오기
@@ -15,24 +14,6 @@ const HeaderRightBox = () => {
   const resetUser = useResetRecoilState(userState);
   const [isLogin, setIsLogin] = useRecoilState(loginState);
   const [isCustomer, setIsCustomer] = useRecoilState(customerState);
-
-  const [isSearchBar, setIsSearchBar] = useState(false);
-  const [searchFilter, setSearchFilter] = useRecoilState(itemListState);
-
-  const searchBarHandler = () => {
-    if (!isSearchBar) setIsSearchBar(true);
-    else setIsSearchBar(false);
-  };
-
-  const searchHandler = (event) => {
-    if (event.key === 'Enter') {
-      setSearchFilter((prevItemList) => ({
-        ...prevItemList,
-        searchKeyword: event.target.value,
-      }));
-      setIsSearchBar(false);
-    }
-  };
 
   const handleLogout = () => {
     removeItem(localstorageKey.auth);
@@ -42,15 +23,8 @@ const HeaderRightBox = () => {
 
   return (
     <S.HeaderRightBox>
-      {isSearchBar && (
-        <S.SearchBarBox>
-          <S.SearchBar type="text" onKeyUp={searchHandler}></S.SearchBar>
-        </S.SearchBarBox>
-      )}
       <S.HeaderBtnsBox>
-        <S.SearchBtnBox onClick={searchBarHandler}>
-          <Icon name="IconSearch" width="20px" />
-        </S.SearchBtnBox>
+        <SearchBar />
         <Link to={isLogin ? '/mypage' : '/login'}>
           <S.MyInfoBtnBox>
             <Icon name="IconUser" width="25px" />
