+++ src/pages/main/Main.jsx
@@ -1,5 +1,3 @@
-import { RecoilRoot } from 'recoil';
-
 import FilterOption from '@/components/main/FilterOption/FilterOption';
 import CarouselBox from '@/components/main/GenreCarousel/CarouselBox';
 import ItemListGrid from '@/components/main/ItemListGrid/ItemListGrid';
@@ -11,16 +9,14 @@ const Main = () => {
   };
 
   return (
-    <RecoilRoot>
-      <S.MainPageWrapper>
-        <S.ToTopBtn onClick={scrollToTop}>TOP</S.ToTopBtn>
-        <S.MainContainer>
-          <CarouselBox />
-          <FilterOption />
-          <ItemListGrid />
-        </S.MainContainer>
-      </S.MainPageWrapper>
-    </RecoilRoot>
+    <S.MainPageWrapper>
+      <S.ToTopBtn onClick={scrollToTop}>TOP</S.ToTopBtn>
+      <S.MainContainer>
+        <CarouselBox />
+        <FilterOption />
+        <ItemListGrid />
+      </S.MainContainer>
+    </S.MainPageWrapper>
   );
 };
